The dimensions of feature/label matrices are now of type uint32.

+++ python/boomer/common/cpp/input_data.h
@@ -21,19 +21,19 @@ class AbstractLabelMatrix {
          * @param numExamples   The number of examples
          * @param numLabels     The number of labels
          */
-        AbstractLabelMatrix(intp numExamples, intp numLabels);
+        AbstractLabelMatrix(uint32 numExamples, uint32 numLabels);
 
         virtual ~AbstractLabelMatrix();
 
         /**
          * The number of examples.
          */
-        intp numExamples_;
+        uint32 numExamples_;
 
         /**
          * The number of labels.
          */
-        intp numLabels_;
+        uint32 numLabels_;
 
 };
 
@@ -48,7 +48,7 @@ class AbstractRandomAccessLabelMatrix : public AbstractLabelMatrix {
          * @param numExamples   The number of examples
          * @param numLabels     The number of labels
          */
-        AbstractRandomAccessLabelMatrix(intp numExamples, intp numLabels);
+        AbstractRandomAccessLabelMatrix(uint32 numExamples, uint32 numLabels);
 
         /**
          * Returns whether a specific label of the example at a given index is relevant or irrelevant.
@@ -78,7 +78,7 @@ class DenseLabelMatrixImpl : public AbstractRandomAccessLabelMatrix {
          * @param y             A pointer to a C-contiguous array of type `uint8`, shape `(numExamples, numLabels)`,
          *                      representing the labels of the training examples
          */
-        DenseLabelMatrixImpl(intp numExamples, intp numLabels, const uint8* y);
+        DenseLabelMatrixImpl(uint32 numExamples, uint32 numLabels, const uint8* y);
 
         ~DenseLabelMatrixImpl();
 
@@ -104,7 +104,7 @@ class DokLabelMatrixImpl : public AbstractRandomAccessLabelMatrix {
          * @param dokMatrixPtr  A shared pointer to an object of type `BinaryDokMatrix`, storing the relevant labels of
          *                      the training examples
          */
-        DokLabelMatrixImpl(intp numExamples, intp numLabels, std::shared_ptr<BinaryDokMatrix> dokMatrixPtr);
+        DokLabelMatrixImpl(uint32 numExamples, uint32 numLabels, std::shared_ptr<BinaryDokMatrix> dokMatrixPtr);
 
         ~DokLabelMatrixImpl();
